requirements.txt.txt
Flask
python-docx
pytesseract
Pillow
PyMuPDF